+++ demetra-dotstat-core/src/main/java/be/nbb/demetra/sdmx/web/SdmxWebBean.java
@@ -1,17 +1,17 @@
 /*
  * Copyright 2015 National Bank of Belgium
- * 
- * Licensed under the EUPL, Version 1.1 or - as soon they will be approved 
+ *
+ * Licensed under the EUPL, Version 1.1 or - as soon they will be approved
  * by the European Commission - subsequent versions of the EUPL (the "Licence");
  * You may not use this work except in compliance with the Licence.
  * You may obtain a copy of the Licence at:
- * 
+ *
  * http://ec.europa.eu/idabc/eupl
- * 
- * Unless required by applicable law or agreed to in writing, software 
+ *
+ * Unless required by applicable law or agreed to in writing, software
  * distributed under the Licence is distributed on an "AS IS" basis,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the Licence for the specific language governing permissions and 
+ * See the Licence for the specific language governing permissions and
  * limitations under the Licence.
  */
 package be.nbb.demetra.sdmx.web;
@@ -20,7 +20,6 @@ import java.time.Duration;
 import java.util.List;
 
 /**
- *
  * @author Philippe Charles
  * @since 2.2.0
  */
@@ -28,6 +27,7 @@ import java.util.List;
 public final class SdmxWebBean {
 
     private String source;
+    private String database;
     private String flow;
     private List<String> dimensions;
     private String labelAttribute;