+++ demetra-dotstat-core/src/main/java/be/nbb/demetra/sdmx/web/SdmxWebParam.java
@@ -16,6 +16,8 @@
  */
 package be.nbb.demetra.sdmx.web;
 
+import static ec.tss.tsproviders.utils.Params.*;
+
 import com.google.common.base.Joiner;
 import com.google.common.base.Splitter;
 import com.google.common.collect.ImmutableList;
@@ -25,13 +27,11 @@ import ec.tss.tsproviders.cube.CubeId;
 import ec.tss.tsproviders.cube.CubeSupport;
 import ec.tss.tsproviders.utils.IConfig;
 import ec.tss.tsproviders.utils.IParam;
-import lombok.NonNull;
-
 import java.time.Duration;
 import java.util.List;
 import java.util.concurrent.TimeUnit;
-
-import static ec.tss.tsproviders.utils.Params.*;
+import lombok.NonNull;
+import sdmxdl.DatabaseRef;
 
 /**
  * @author Philippe Charles
@@ -44,14 +44,19 @@ interface SdmxWebParam extends IParam<DataSource, SdmxWebBean> {
 
     final class V1 implements SdmxWebParam {
 
-        private final Splitter dimensionSplitter = Splitter.on(',').trimResults().omitEmptyStrings();
+        private final Splitter dimensionSplitter =
+                Splitter.on(',').trimResults().omitEmptyStrings();
         private final Joiner dimensionJoiner = Joiner.on(',');
 
         private final IParam<DataSource, String> dbName = onString("", "dbName");
+        private final IParam<DataSource, String> database =
+                onString(DatabaseRef.NO_DATABASE.toString(), "db");
         private final IParam<DataSource, String> flowRef = onString("", "tableName");
-        private final IParam<DataSource, List<String>> dimensionIds = onStringList(ImmutableList.of(), "dimColumns", dimensionSplitter, dimensionJoiner);
+        private final IParam<DataSource, List<String>> dimensionIds =
+                onStringList(ImmutableList.of(), "dimColumns", dimensionSplitter, dimensionJoiner);
         private final IParam<DataSource, String> labelAttribute = onString("", "l");
-        private final IParam<DataSource, Long> cacheTtl = onLong(TimeUnit.MILLISECONDS.convert(5, TimeUnit.MINUTES), "cacheTtl");
+        private final IParam<DataSource, Long> cacheTtl =
+                onLong(TimeUnit.MILLISECONDS.convert(5, TimeUnit.MINUTES), "cacheTtl");
         private final IParam<DataSource, Integer> cacheDepth = onInteger(1, "cacheDepth");
 
         @Override
@@ -63,6 +68,7 @@ interface SdmxWebParam extends IParam<DataSource, SdmxWebBean> {
         public @NonNull SdmxWebBean defaultValue() {
             SdmxWebBean result = new SdmxWebBean();
             result.setSource(dbName.defaultValue());
+            result.setDatabase(database.defaultValue());
             result.setFlow(flowRef.defaultValue());
             result.setDimensions(dimensionIds.defaultValue());
             result.setLabelAttribute(labelAttribute.defaultValue());
@@ -75,6 +81,7 @@ interface SdmxWebParam extends IParam<DataSource, SdmxWebBean> {
         public @NonNull SdmxWebBean get(@NonNull DataSource dataSource) {
             SdmxWebBean result = new SdmxWebBean();
             result.setSource(dbName.get(dataSource));
+            result.setDatabase(database.get(dataSource));
             result.setFlow(flowRef.get(dataSource));
             result.setDimensions(dimensionIds.get(dataSource));
             result.setLabelAttribute(labelAttribute.get(dataSource));
@@ -87,6 +94,7 @@ interface SdmxWebParam extends IParam<DataSource, SdmxWebBean> {
         public void set(IConfig.@NonNull Builder<?, DataSource> builder, SdmxWebBean value) {
             if (value != null) {
                 dbName.set(builder, value.getSource());
+                database.set(builder, value.getDatabase());
                 flowRef.set(builder, value.getFlow());
                 dimensionIds.set(builder, value.getDimensions());
                 labelAttribute.set(builder, value.getLabelAttribute());